moves all the seeding files for postgresql into one folder and set the csv generation to ten million records
