--- server/exampleCSV.csv
@@ -1,2 +1,0 @@
-3|jimmmyjohns|something1
-4|jackieO|something2
